Fix #8 Create Pigosat-specific types for literals, clauses, etc.
This makes it easier for users of the API to remember the relationships among
the different argument and return-value types from the different Pigosat
methods. Further, it makes maintenance easier: e.g., we can change the literal
type from int32 to int64 just by changing the `type Literal int32` line.