Partially revert change to Pigosat.Seconds test from 2889cc3f539
2889cc3f539 turns off the test of Pigosat.Seconds by switching the consequence
of failing the test from Error to Log. This commit turns the test back on
(reinstating Error). However, some of the smaller tests in the formulaTests
table take less time than the resolution of the timer and thus come out to zero.
This causes frequent failures of this test. Therefore, I have relaxed the test
to allow times of zero.